Added data normalization to TLearnPattern class
- Restore value functions added to TLearnPattern class
- Update() and PropagateNetBatchBackward() functions in in TNeuroNet
class are template now: can accept any vector type double*, TPatternRow,
vector<double>, etc.
- Normalization range constants added for each activation function in
TNeuroNet class (utilized by TLearnPattern)

+++ lib/tdatasource.cpp
@@ -56,12 +56,12 @@ int TDataSource::max(string colName) throw(invalid_argument)
     return maxByColName[colName];
 }
 
-void TDataSource::calcColumnMinMax(int col)
+void TDataSource::calcColumnMinMax(uint col)
 {
     if(rowCount() <= 0) return;
 
     double min = data(0, col), max = data(0, col);
-    for(int i = 1; i < rowCount(); i++)
+    for(size_t i = 1; i < rowCount(); i++)
     {
         if(min > data(i, col)) min = data(i, col);
         if(max < data(i, col)) max = data(i, col);
